dji sim consolidation
